feat: Create a dummy api server
In this commit, i created a dummy Api/server for getting dummy stories. The stories gotten
from this api is then used to create a list of stories from users.

This api endpoint is temporary and should be replaced with the endpoint provided by
the backend devs of the coordinated-story-telling team