feat: Add Storiesdetails component
In this commit, I added a new component <Storiesdetails /> for displaying the details
about individual stories. Details rendered to the screen are:
the author ( i.e owner of the story)
the title of the story
and the story itself

A styled component was also used in adding custom styles to the Storiesdetails card.